bug fix and feature update for logs (#97)
* bug fix and feature update for logs

added a toggle for logging. this way we don't saturate the logs and we have fine control on the process.
changed the logic around checking for store ID and isolate the check for storeID instead of having 2 checks.
Also fixed a decimal rounding issue.

* Update class-helloextend-protection-logger.php

added an initialize function to make sure the settings is never undefined.

* Update class-helloextend-global.php

added log_enabled and debug_log_enabled to the compact function

* Update class-helloextend-global.php

removed unused variable

* Update class-helloextend-global.php

removed a circular reference

* Refactored frontend test to use in-stock product

---------

Co-authored-by: Alex Smith <alex.smith@extend.com>